+++ packages/analytics-js/src/components/capabilitiesManager/detection/index.ts
@@ -1,5 +1,5 @@
 export { detectAdBlockers } from './adBlockers';
-export { isBrowser, isNode, hasCrypto, hasUAClientHints, hasBeacon, isIE11 } from './browser';
+export { isBrowser, isNode, hasCrypto, hasUAClientHints, hasBeacon } from './browser';
 export { getUserAgentClientHint } from './clientHint';
 export { isDatasetAvailable, legacyJSEngineRequiredPolyfills, isLegacyJSEngine } from './dom';
 export { getScreenDetails } from './screen';