+++ packages/analytics-js-common/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.25.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.24.0...@rudderstack/analytics-js-common@3.25.0) (2025-09-15)
+
+
+### Features
+
+* optimize page leave events for bfcache ([#2458](https://github.com/rudderlabs/rudder-sdk-js/issues/2458)) ([48c1f20](https://github.com/rudderlabs/rudder-sdk-js/commit/48c1f208efa0edce88fe9bedaafd87511fb0a365))
+* reset api with options ([#2445](https://github.com/rudderlabs/rudder-sdk-js/issues/2445)) ([f739030](https://github.com/rudderlabs/rudder-sdk-js/commit/f73903092e20d45a6cf047cc03fe16f5290fb557))
+
 ## [3.24.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.23.0...@rudderstack/analytics-js-common@3.24.0) (2025-09-01)