+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.13.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.13.0...@rudderstack/analytics-js-plugins@3.13.1) (2025-09-15)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.25.0`
+* `@rudderstack/analytics-js-cookies` updated to version `0.5.3`
 ## [3.13.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.12.1...@rudderstack/analytics-js-plugins@3.13.0) (2025-09-01)
 
 ### Dependency Updates